Extract batch delete logic to function on NSEntityDescription
